refactor(agent): consolidate orchestration types into run-worker
Move all type definitions from orchestration.ts to run-worker.ts:
- OrchestratorConfig, OrchestrationResult, OrchestrationState
- ToolExecutor, SessionMode, OrchestratorRunInput

Remove Orchestrator namespace (was thin facade delegating to RunWorker.run).
Update all 8 importing files to import from run-worker instead.
Update loop/index.ts barrel export to re-export from run-worker.

This eliminates the compatibility facade and consolidates orchestration
types with their primary implementation.
