+++ packages/agent/src/loop/run-worker.ts
@@ -15,21 +15,56 @@ import { Observability } from "./observability";
 import { AuditLog } from "./audit";
 import { DeadLetterQueue } from "./dlq";
 import { SummaryDelivery } from "./summary";
-import type {
-  OrchestratorConfig,
-  OrchestrationResult,
-  OrchestratorRunInput,
-  SessionMode,
-  ToolExecutor,
-} from "./orchestration";
+
+type PermissionLevel = "ask" | "notify" | "deny";
+
+export type SessionMode = "ephemeral" | "persistent" | "reuse";
+
+export interface OrchestratorConfig {
+  taskId: string;
+  runId: string;
+  maxRetries: number;
+  sessionMode?: SessionMode;
+  sessionId?: string;
+  maxSubagentDepth?: number;
+  currentDepth?: number;
+  insideDelegation?: boolean;
+}
+
+export interface OrchestrationResult {
+  success: boolean;
+  summary: string;
+  error: string;
+}
+
+export interface OrchestrationState {
+  attempt: number;
+  turns: number;
+  toolCalls: number;
+  lastError: string;
+}
+
+export interface ToolExecutor {
+  execute(calls: ToolCall[]): Promise<ToolResult[]>;
+}
+
+export interface OrchestratorRunInput {
+  llm: {
+    run(input: Record<string, unknown>, sink: Sink): Promise<RunOutcome>;
+  };
+  input: Record<string, unknown>;
+  toolExecutor?: ToolExecutor;
+  permission?: {
+    agentPolicy?: PermissionLevel;
+    systemDefault?: PermissionLevel;
+  };
+}
 
 type RetryReason =
   | "timeout"
   | "tool_error"
   | "transient_error"
   | "validation_error";
-
-type PermissionLevel = "ask" | "notify" | "deny";
 
 const DEFAULT_PERMISSION: PermissionLevel = "notify";
